feat: add non-form inputs
