+++ src/lib/shared/components/ui/index.ts
@@ -8,11 +8,15 @@ import ContextMenuContent from './context-menu/ContextMenuContent.svelte';
 import ContextMenuItem from './context-menu/ContextMenuItem.svelte';
 import ContextMenuSeparator from './context-menu/ContextMenuSeparator.svelte';
 import ContextMenuTrigger from './context-menu/ContextMenuTrigger.svelte';
+import DateInput from './DateInput.svelte';
 // drag drop
 import { DragManager } from './drag-drop/drag-manager.js';
 import Draggable from './drag-drop/Draggable.svelte';
 import Dropzone from './drag-drop/Dropzone.svelte';
+import Input from './Input.svelte';
 import SearchBar from './search/SearchBar.svelte';
+import Select from './Select.svelte';
+import TextArea from './TextArea.svelte';
 
 export {
   Choice,
@@ -23,12 +27,16 @@ export {
   ContextMenuItem,
   ContextMenuSeparator,
   ContextMenuTrigger,
+  DateInput,
   Draggable,
 
   // drag drop
   DragManager,
   Dropzone,
+  Input,
   SearchBar,
+  Select,
+  TextArea,
   WeekdayChoice,
   WeekdayChoiceMulti
 };
